Bump Ruby and disable HttpPositionalArguments
All of the projects are now migrated to ruby 2.3.3 so let's bump the
version in the rubocop config so we don't get errors for new features of
the language.

Rails/HttpPositionalArguments option works only for Rails 5 applications
(https://github.com/bbatsov/rubocop/issues/3629). So I propose to disable
this cop for the moment.